experimental/requirements.txt
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
Pillow>=10.0.0
pandas>=2.0.0

